Fixed "Window" Error in ProfileSection, Modified SideBar, added Card.
ProfileSection: Was presenting an error in the form of a missing "Window" object, this was because the application was attempting to perform SSR but ParticlesBG needs access to the window object. Corrected by forcing SSR: False when importing ProfileSection in SideBar.

SideBar: Modified CCS properties to fill the entire side of the page with the SideBar. Also made the selected section transparent to better indicate which section is selected.

Card: Added a card component with paralax on mouse hover for use to display projects.

To-Do:
- Style the card component, should be generic with props being passed to fill content. Required sections include title, description, image, href.